reorganizando codigo
