+++ .gitignore
@@ -25,5 +25,5 @@ node_modules/
 venv/
 media/
 downloads/
-/src/data/config.yaml
-/src/data/save.json
+/data/config.yaml
+/data/save.json
