+++ series.yaml
